Fixes #6
Change `<link src="...">` to `<link href="...">`. Also, fix the
alignment of the search bar by removing an unnecessary div and adding a
property to its style rule.

Also, change the jquery snippet which highlights the proper nav page
indicator to compare each link's `href` with the current location MINUS
the bookmark part of the url.