Add matchers for individual form elements
Allow individual inputs to be matched by regex. By default, inputs of a
null or empty-string value are considered invalid, but if a matcher is
specified for the input in the `matchers` parameter, it will be used as
a regular expression to test the corresponding input on whichever form
is being validated.

Remove support for adding validation to multiple forms in one call
(this is a rare case and disallows any other parameters from being
passed into the validation function).

Change the `Ghetto` global container to be a `var` instead of a `const`
(consts are block-scoped and throw errors upon re-definition
:frowning:).